+++ app/src/main/java/ua/com/radiokot/photoprism/api/photos/model/PhotoPrismMergedPhoto.kt
@@ -3,6 +3,9 @@ package ua.com.radiokot.photoprism.api.photos.model
 import com.fasterxml.jackson.annotation.JsonCreator
 import com.fasterxml.jackson.annotation.JsonProperty
 
+/**
+ * [photo_results.go](https://github.com/photoprism/photoprism/blob/47e9b20929e44bd63d32cb50a393371d24bcaff8/internal/search/photos_results.go#L15)
+ */
 data class PhotoPrismMergedPhoto
 @JsonCreator
 constructor(
@@ -28,11 +31,11 @@ constructor(
     val mainFile: File?
         // https://github.com/photoprism/photoprism/blob/2f9792e5411f6bb47a84b638dfc42d51b7790853/frontend/src/model/photo.js#L520
         get() = files.find {
-            it.primary
+            it.primary == true
         } ?: files.find {
             it.fileType == "jpg" || it.fileType == "png"
         } ?: files.find {
-            !it.sidecar
+            it.sidecar == false
         }
 
     val videoFile: File?
@@ -42,7 +45,7 @@ constructor(
         } ?: files.find {
             it.fileType == "mp4"
         } ?: files.find {
-            it.video
+            it.video == true
         } ?: animatedFile
 
     private val animatedFile: File?
@@ -51,7 +54,11 @@ constructor(
             it.fileType == "gif" || it.duration != null || it.frames != null
         }
 
+    /**
+     * [file.go](https://github.com/photoprism/photoprism/blob/47e9b20929e44bd63d32cb50a393371d24bcaff8/internal/entity/file.go#L40)
+     */
     class File(
+        // TODO Clarify whether the hash is actually optional.
         @JsonProperty("Hash")
         val hash: String,
         @JsonProperty("UID")
@@ -61,24 +68,24 @@ constructor(
         @JsonProperty("Name")
         val name: String,
         @JsonProperty("Mime")
-        val mime: String,
+        val mime: String?,
         @JsonProperty("FileType")
-        val fileType: String,
+        val fileType: String?,
         @JsonProperty("Size")
-        val size: Long,
+        val size: Long?,
         @JsonProperty("Duration")
         val duration: Long?,
         @JsonProperty("Frames")
         val frames: Int?,
         @JsonProperty("Primary")
-        val primary: Boolean,
+        val primary: Boolean?,
         @JsonProperty("Root")
-        val root: String,
+        val root: String?,
         @JsonProperty("Video")
-        val video: Boolean,
+        val video: Boolean?,
         @JsonProperty("Codec")
         val codec: String?,
         @JsonProperty("Sidecar")
-        val sidecar: Boolean,
+        val sidecar: Boolean?,
     )
 }